TW: bug in cleanup.sh - deleted everything!

+++ forecast_scripts/hindcast/TP4a0.12/ice_only/cleanup.sh
@@ -2,7 +2,7 @@ source $SWARP_ROUTINES/source_files/hex_vars.src
 refno=5
 
 E0_start=$((1+1$refno))
-E0_stop=65
+E0_stop=99
 P=$TP4_REALTIME
 
 for E0 in `seq $E0_start $E0_stop`
@@ -13,8 +13,17 @@ do
    bdir=$P/Build_V2.2.12_X$X  # new Build directory (has hycom executable compiled)
    # echo $E $X; continue
 
+   if [ ! -d $xdir ] && [ ! -d $bdir ]
+   then
+      echo "($xdir,$bdir) not present"
+      continue
+   fi
+   # echo $E $X; continue
+
    # delete expt and build dir's
+   echo rm -rf $xdir $bdir
    rm -rf $xdir $bdir
+   # exit
 
    # clean force etc
    cd $P/force/rivers
